privac

+++ lib/constants.dart
@@ -25,5 +25,5 @@ class Constants {
 
 class UserData {
   static int dollar = 500;
-  static int accountBalance = 11500;
+  static double accountBalance = 11500;
 }
